feat(Guild): move `GuildUser` here
