+++ khl/__init__.py
@@ -9,9 +9,9 @@ from .client import Client
 
 # concepts
 from .role import Role
-from .user import User, GuildUser
+from .user import User
 from .channel import Channel, PublicTextChannel, PrivateChannel, PublicChannel
-from .guild import Guild
+from .guild import Guild, GuildUser
 from .context import Context
 from .message import RawMessage, Message, PublicMessage, PrivateMessage, Event
 # extensions
